fix threshold index type

+++ taichi/dnnc.py
@@ -331,7 +331,7 @@ class DNNC(object):
                                                                 self.tokenizer, self.train_data, self.train_label_ids, 
                                                                 unique_labels, self.device)
         # compute index to print per threshold entered
-        threshold_index = config.threshold * 100 
+        threshold_index = int(config.threshold * 100) 
         logger.info(f"in-domain eval at {config.threshold} threshold: {res_indomain[theshold_index]}")
         res_ood, prob_ood = self._evaluation_ood_recall(model, language, self.ood_test_data, self.tokenizer, self.train_data, 
                                                 unique_labels, self.device)
